requirements.txt
streamlit
azure-storage-blob
pyodbc
dotenv
